build: pass -DUSE_ROBIN_HOOD_HASHING to compilers
The file vk_layer_data.h protects the include of robin_hood.h:

    #ifdef USE_ROBIN_HOOD_HASHING
    #include "robin_hood.h"
    #else
    #include <unordered_set>
    #endif

But the -DUSE_ROBIN_HOOD_HASHING=ON|OFF passed to CMake does nothing
to pass the same definition to the compilers.  This change ensures
that the compilers receive -DUSE_ROBIN_HOOD_HASHING if the CMake
switch is on, or nothing if the switch is off.